+++ example/config.go
@@ -12,7 +12,7 @@ func GetConfig() {
 	brokerUrl := config.GetString("worker", "broker_url")
 	fmt.Println(brokerUrl)
 
-	concurrency := config.GetInt("worker", "concurrency")
+	concurrency := config.GetInt64("worker", "concurrency")
 	fmt.Println(concurrency)
 
 	boole := config.GetBoole("worker", "boole")
